Update all app
