+++ src/components/MenuHeader/index.js
@@ -1,0 +1,13 @@
+import Menu from "./Menu";
+import Navbar from "./Navbar";
+
+const MenuHeader = () => {
+  return (
+    <>
+      <Menu />
+      <Navbar />
+    </>
+  );
+};
+
+export default MenuHeader;
