+++ src/components/PokemonCard/index.js
@@ -1,8 +1,9 @@
 import { useState } from "react";
-
-import style from "./style.module.css";
+import cn from "classnames";
 
 import bgImg from "./assets/card-back-side.jpg";
+
+import style from "./style.module.css";
 
 function PokemonCard({ name, img, id, type, values }) {
   const [isActive, setActive] = useState(false);
@@ -13,39 +14,35 @@ function PokemonCard({ name, img, id, type, values }) {
 
   return (
     <div className={style.root} onClick={handleClick}>
-      <div className={`${style.pokemonCard} ${isActive ? style.active : ""}`}>
-        <div className={style.cardFront}>
-          <div className={`${style.wrap} ${style.front}`}>
-            <div className={`${style.pokemon} ${style[type]}`}>
-              <div className={style.values}>
-                <div className={`${style.count} ${style.top}`}>
-                  {values.top}
-                </div>
-                <div className={`${style.count} ${style.right}`}>
+      <div className={cn(style.pokemonCard, { [style.active]: isActive })}>
+        <div className={cn(style.cardFront)}>
+          <div className={cn(style.wrap, style.front)}>
+            <div className={cn(style.pokemon, style[type])}>
+              <div className={cn(style.values)}>
+                <div className={cn(style.count, style.top)}>{values.top}</div>
+                <div className={cn(style.count, style.right)}>
                   {values.right}
                 </div>
-                <div className={`${style.count} ${style.bottom}`}>
+                <div className={cn(style.count, style.bottom)}>
                   {values.bottom}
                 </div>
-                <div className={`${style.count} ${style.left}`}>
-                  {values.left}
-                </div>
+                <div className={cn(style.count, style.left)}>{values.left}</div>
               </div>
-              <div className={style.imgContainer}>
+              <div className={cn(style.imgContainer)}>
                 <img src={img} alt={name} />
               </div>
-              <div className={style.info}>
-                <span className={style.number}>{id}</span>
-                <h3 className={style.name}>{name}</h3>
-                <small className={style.type}>
+              <div className={cn(style.info)}>
+                <span className={cn(style.number)}>{id}</span>
+                <h3 className={cn(style.name)}>{name}</h3>
+                <small className={cn(style.type)}>
                   <span>{type}</span>
                 </small>
               </div>
             </div>
           </div>
         </div>
-        <div className={style.cardBack}>
-          <div className={`${style.wrap} ${style.back}`}>
+        <div className={cn(style.cardBack)}>
+          <div className={cn(style.wrap, style.back)}>
             <img src={bgImg} alt="Сard Backed" />
           </div>
         </div>
