+++ src/routes/Game/index.js
@@ -1,0 +1,9 @@
+const GamePage = ({ onChangePage2 }) => {
+  return (
+    <div>
+      <button onClick={onChangePage2}>Home page</button>
+    </div>
+  );
+};
+
+export default GamePage;
